requirements.txt
# AWS SDK
boto3==1.28.38
aws-cdk-lib==2.94.0
constructs>=10.0.0
aws-cdk.aws-apigatewayv2-alpha>=2.94.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha>=2.94.0a0

# Application dependencies
pydantic==2.3.0
python-json-logger==2.0.7

# Development and testing
pytest==7.4.0
moto==4.2.5
pytest-mock==3.11.1